Check the linux host if the child prefix already exists
- resolves part of #73 and prevents an exit
