+++ internal/aircrew/darwin-utils.go
@@ -1,0 +1,8 @@
+//go:build darwin
+
+package aircrew
+
+// routeExists currently only used for darwin build purposes
+func routeExists(s string) bool {
+	return false
+}
