+++ internal/aircrew/utils.go
@@ -12,6 +12,7 @@ import (
 	"strings"
 
 	log "github.com/sirupsen/logrus"
+	"github.com/vishvananda/netlink"
 )
 
 // supported OS types
@@ -196,4 +197,34 @@ func ParseIPNet(s string) (*net.IPNet, error) {
 		return nil, err
 	}
 	return &net.IPNet{IP: ip, Mask: ipNet.Mask}, nil
+}
+
+// deleteIface checks to see if  is an interface exists and deletes it
+func linkExists(ifaceName string) bool {
+	if _, err := netlink.LinkByName(ifaceName); err != nil {
+		return false
+	}
+	return true
+}
+
+// delLink deletes the link and assumes it exists
+func delLink(ifaceName string) error {
+	if link, err := netlink.LinkByName(ifaceName); err == nil {
+		if err = netlink.LinkDel(link); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// IfaceUP brings a specified netlink interface up
+func IfaceUP(ifaceName string) error {
+	link, err := netlink.LinkByName(ifaceName)
+	if err != nil {
+		return fmt.Errorf("failed to find the specified interface %s: %v", ifaceName, err)
+	}
+	if err := netlink.LinkSetUp(link); err != nil {
+		return fmt.Errorf("failed to enable the specified interface %s: %v", ifaceName, err)
+	}
+	return nil
 }
